requirements.txt
docling
python-dotenv>=1.0
